refactor: extract Storage class into storage.py (Phase 2/3)
Move the entire Storage class (~1550 lines) from api.py into its own
module, storage.py.  This is the largest single extraction in the
refactoring plan — Storage contains all PostgreSQL connection pooling,
table initialization, and every CRUD method.

Changes:
- Created storage.py with Storage class and hash_api_key helper
- api.py now imports: from storage import Storage, hash_api_key
- Removed unused imports from api.py (psycopg2, urlparse, hashlib)
- All 106 existing tests pass unchanged

api.py: 3877 → 2314 lines (-40%)
storage.py: 1576 lines (new)

Part of #66
